Don't use AS.on_load for something else than on_load hook
ActiveSupport.on_load is not just a way to add an additional block of code into a Ruby method.
It's a mechanism for adding load hooks which is needed to implement lazy loading in ActiveSupport.

And, in this particular after_established case, we don't actually need any special mechanism. Let's just simply call the logic inside the AMC.