Fix bypassing uniqueness validators
When `Validator` is iterating over a callback chain and skips unique
validations, it shifts the index whenever it deletes such item from the
collection. If there are several successive uniqueness validators, it
omits the one following the deleted one in the chain. That prevents
from inserting data in such model when running with `:validate` option.

Do not mutate the target chain inside the enumerator and iterate over a
copy instead.

Relates to https://github.com/zdennis/activerecord-import/pull/301.
Follow-up for https://github.com/zdennis/activerecord-import/pull/410.